xml import should be able to import files/images again, commands able to recover from 'could not resolve host' errors, controllers should correctly ensure directories exist

+++ src/ODR/AdminBundle/Command/CSVExportStartCommand.php
@@ -43,7 +43,6 @@ class CSVExportStartCommand extends ContainerAwareCommand
         // Only need to load these once...
         $container = $this->getContainer();
         $logger = $container->get('logger');
-        $router = $container->get('router');
         $pheanstalk = $container->get('pheanstalk');
 
         // Run command until manually stopped
@@ -62,7 +61,6 @@ class CSVExportStartCommand extends ContainerAwareCommand
                 $current_time = new \DateTime();
                 $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );                
                 $output->writeln($str);
-//                $logger->info('CSVExportStartCommand.php: '.$str);
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
@@ -96,7 +94,13 @@ $output->writeln($data->url);
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
@@ -121,11 +125,24 @@ $output->writeln($data->url);
 
             }
             catch (\Exception $e) {
-                $output->writeln($e->getMessage());
-                $logger->err('CSVExportStartCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('CSVExportStartCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue doesn't hang, in theory
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+
+                    $logger->err('CSVExportStartCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
